requirements.txt
accelerate==1.9.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
bitsandbytes==0.46.1
blis==1.3.0
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools==5.5.2
catalogue==2.0.10
certifi @ file:///C:/b/abs_43o9lm92p9/croot/certifi_1750863864870/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cymem==2.0.11
debugpy==1.8.16
decorator==5.2.1
decord==0.6.0
Deprecated @ file:///C:/Windows/TEMP/abs_4flgviyh6m/croots/recipe/deprecated_1659726430166/work
docstring_parser==0.17.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
executing==2.2.0
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
fsspec @ file:///C:/b/abs_ddvco8a9tq/croot/fsspec_1750682718626/work
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.105.0
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.27.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
icecream==2.1.5
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///C:/b/abs_9d0lp1ysmt/croot/importlib_metadata-suite_1732633510972/work
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keybert==0.9.0
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
langcodes==3.5.0
language_data==1.3.0
lap==0.5.12
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numexpr @ file:///C:/b/abs_4e0h6i5ip2/croot/numexpr_1752521729801/work
numpy==1.26.4
opencv-python==4.12.0.88
opentelemetry-api @ file:///C:/b/abs_f1q379y4o0/croot/opentelemetry-api_1743171676016/work
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp310-cp310-win_amd64.whl#sha256=6fc0a0523ea4ec3f23a2b790c738c857942b4c579479f8012ac8148d675b8975
parso==0.8.4
pillow @ file:///C:/b/abs_b06nvnir5m/croot/pillow_1752524543576/work
platformdirs==4.3.8
preshed==3.0.10
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt6==6.7.1
PyQt6_sip @ file:///C:/b/abs_f47pmp6xmn/croot/pyqt-split_1744804513582/work/pyqt_sip
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pytextrank==3.3.0
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytz @ file:///C:/b/abs_f8wdzeix0n/croot/pytz_1752135878094/work
pywin32==311
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq==27.0.2
regex==2024.11.6
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
rich==14.1.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.13.1
sentence-transformers==5.0.0
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
sip @ file:///C:/b/abs_5cto136kse/croot/sip_1738856220313/work
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tenacity==8.5.0
thinc==8.3.6
threadpoolctl==3.6.0
timm==1.0.17
tokenizers==0.21.2
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado @ file:///C:/b/abs_7dzpc171lf/croot/tornado_1748956950306/work
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
ultralytics==8.3.168
ultralytics-thop==2.0.14
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
uritemplate==4.2.0
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wrapt @ file:///C:/b/abs_68gzx33lzh/croot/wrapt_1736543196433/work
zipp @ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work
